refactor: Dissolve util/ into the layers that own each module
Types and helpers live with something real, not in grab-bag modules:

- util/prompt.ts -> interact/prompt.ts: the prompting primitive is the
  foundation of the interaction layer.
- util/with-loading.ts -> output/with-loading.ts: a stderr spinner
  gated on the output format is an output concern.
- util/ellipsis.ts -> render/text.ts: text truncation for display
  belongs to the presentation layer (markdown helpers join it later).
