+++ src/lib/interact/interact-for-custom-metadata.test.ts
@@ -3,12 +3,12 @@ import { beforeEach, expect, test, vi } from 'vitest'
 import { interactForCustomMetadata } from './interact-for-custom-metadata.js'
 import { createMemoryOutput } from '../output/create-memory-output.js'
 import { setOutput } from '../output/get-output.js'
-import type * as PromptModule from '../util/prompt.js'
-import { promptSelect, promptText } from '../util/prompt.js'
+import type * as PromptModule from './prompt.js'
+import { promptSelect, promptText } from './prompt.js'
 
 // Only the prompts themselves are replaced, so the real PromptCancelledError
 // and withBackHint are used, as they are in production.
-vi.mock('../util/prompt.js', async (importOriginal) => ({
+vi.mock('./prompt.js', async (importOriginal) => ({
   ...(await importOriginal<typeof PromptModule>()),
   promptText: vi.fn(),
   promptNumber: vi.fn(),
