+++ src/lib/render/text.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from 'vitest'
 
-import { ellipsis } from './ellipsis.js'
+import { ellipsis } from './text.js'
 
 test('ellipsis: truncates only when over the limit', () => {
   expect(ellipsis('seam', 10)).toBe('seam')
