+++ src/lib/render/text.ts
